+++ src/components/common/ControlsWrapper/style.module.scss
@@ -1,5 +1,6 @@
 .controls-wrapper {
   display: flex;
   align-items: center;
-  margin: 10px 0;
+  margin-top: 10px;
+  margin-bottom: 10px;
 }
